requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0

numpy==1.26.4
PyYAML==6.0.3
pydantic==2.13.3
pydantic-settings==2.14.0
nltk==3.9.4
sentence-transformers==5.4.1
faiss-cpu==1.13.2
openai==2.33.0
python-dotenv==1.2.2